Setup directory and templates for a new research study

This function can be used to automatically setup your study directory by creating folders and template scripts

create_project(
  script_templates = FALSE,
  mainscript = FALSE,
  rawscript = FALSE,
  scorescript = FALSE,
  mergescript = FALSE,
  analysisscript = FALSE,
  study_doc = FALSE,
  data_raw = FALSE,
  data_scored = FALSE,
  documents = FALSE,
  results = FALSE,
  tasks = FALSE,
  sessions = NULL,
  manuscript = FALSE,
  presentations = FALSE,
  figures = FALSE,
  other = c(),
  standard_project = FALSE,
  path = "."
)

Arguments

script_templates
Logical. Download all script templates? default = FALSE

mainscript
Logical. Download a main script template? default = TRUE

rawscript
Logical. Download a raw script template? default = FALSE

scorescript
Logical. Download a score script template? default = FALSE

mergescript
Logical. Download a merge script template? default = FALSE

analysisscript
Logical. Download template to do data analysis (.Rmd)? default = FALSE

study_doc
Logical. Download a study documentation template? default = FALSE

data_raw
Logical. Create a raw data directory? default = FALSE

data_scored
Logical. Create a scored data directory? default = FALSE

documents
Logical. Create documents directory? default = FALSE

results
Logical. Create results directory? default = FALSE

tasks
Logical. Create tasks directory? default = FALSE

sessions
Numeric. How many sessions will the study have?

manuscript
Logical. Create manuscript directory? default = FALSE

presentations
Logical. Create presentations directory? default = FALSE

figures
Logical. Create figures directory? default = FALSE

other
List of other directories you want to create

standard_project
Logical. Create a standard project directory? default = FALSE

path
String. Home directory file path