Check and Remove Duplicate IDs

Checks and removes duplicate subject ids that can happen when misentered during task administration

duplicates_check(
  x,
  unique_id = "Subject",
  date_time = c("SessionDate", "SessionTime"),
  n = 1,
  remove = TRUE,
  keep_by = c("none", "first date_time", "least missing"),
  save_as = NULL,
  id = "Subject",
  unique = c("SessionDate", "SessionTime")
)

Arguments

x

dataframe

unique_id

Column name(s) that should be uniquely identified and will be used to check for duplicate id's grouped by these columns (e.g., "Subject", c("Subject, Session))

date_time

Column names that uniquely identify a testing administration. (e.g., "Date", "Time"). Can also be: "none" or NULL

n

Number of unique id's expected (default: 1).

remove

logical. Remove duplicate ids from data? (default: TRUE)

keep_by

Which duplicate id should be kept? options: "none", "first date_time", "least missing"

save_as

Folder path and file name to output the duplicate ID's

id

deprecated. Use unique_id instead.

unique

deprecated. Use date_time instead.