Bind Multiple Files

Binds multiple files in the same directory into a single data frame. Bind can occur by "rows" or "columns".

files_bind(
  path = "",
  pattern = "",
  delim = ",",
  bind = "rows",
  output_file = "",
  source = ""
)

Arguments

path

Folder location of files to be binded

pattern

Pattern to identify files to be binded

delim

Delimiter used in files. Passed onto `readr::read_delim()`

bind

Bind "rows" or "columns"?

output_file

File name and path to be saved to.

source

If data file comes from a specific software specify it here for better support. Currently supports "eprime".